# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "board Rest API",
        "description": "board API",
        "contact": {
            "email": "chkim@dtwocorp.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "",
            "description": ""
        }
    ],
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "/login"
                ],
                "summary": "로그인",
                "description": "로그인",
                "operationId": "bae68d1c55329ecb6e41f48285840c8f",
                "requestBody": {
                    "description": "Login data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User logged in successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "성공"
                                        },
                                        "data": {
                                            "properties": {
                                                "access_token_name": {
                                                    "type": "string",
                                                    "example": "X-ACCESS-TOKEN"
                                                },
                                                "access_token": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "/logout"
                ],
                "summary": "로그아웃",
                "description": "로그아웃",
                "operationId": "c5cdd87fdbf9d0f95dede875bd023dec",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "성공"
                                        },
                                        "data": {
                                            "properties": {
                                                "access_token_name": {
                                                    "type": "string",
                                                    "example": "X-ACCESS-TOKEN"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}